feat: upgrade axios library
