+++ frontendv3/package.json
@@ -9,10 +9,10 @@
   },
   "dependencies": {
     "awesome-cron": "^0.0.14",
-    "axios": "^0.27.2",
+    "axios": "^1.7.7",
     "chart.js": "^4.2.1",
-    "naive-ui": "^2.32.2",
-    "pnpm": "^8.5.1",
+    "frontendv3": "file:",
+    "pnpm": "^8.15.8",
     "qs": "^6.10.5",
     "vue": "^3.2.25",
     "vue-chartjs": "^5.2.0",
@@ -23,6 +23,7 @@
     "@types/qs": "^6.9.7",
     "@vicons/ionicons5": "^0.12.0",
     "@vitejs/plugin-vue": "^3.1.0",
+    "naive-ui": "^2.40.1",
     "seemly": "^0.3.6",
     "typescript": "^4.5.4",
     "unplugin-vue-components": "^0.22.7",
